backend/requirements.txt
Django==3.2.10
djangorestframework==3.12.4
django-cors-headers==3.10.0
flake8==3.9.2
gunicorn==20.1.0